pyproject.toml
[tool.poetry]
name = "shop-api"
version = "0.1.0"
description = ""
authors = ["Piotr Stempkowski <p.stempkowski@futuremind.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
ipython = "^8.13.2"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
